pydantic/deep_research_agent/requirements.txt
pydantic-ai
gradio
python-dotenv
openai
duckduckgo-search
beautifulsoup4
requests
html2text
